Refactoring add-filter-button-on-bookmark-page
