+++ src/dmm/books/add-filter-button-on-bookmark-page/createActiveButton.ts
@@ -1,0 +1,12 @@
+export const createActiveButton = (text: string): HTMLLIElement => {
+  const label = document.createElement('span')
+  label.style.paddingLeft = '8px'
+  label.style.paddingRight = '8px'
+  label.textContent = text
+
+  const button = document.createElement('li')
+  button.classList.add('current')
+  button.style.width = 'auto'
+  button.append(label)
+  return button
+}
