+++ src/dmm/books/add-filter-button-on-bookmark-page/createButtonList.ts
@@ -1,8 +1,6 @@
-// TODO: リファクタリング
-
 import { createAllButton } from './createAllButton'
 import { createDiscountedButton } from './createDiscountedButton'
-import { FilterType } from './FilterType'
+import type { FilterType } from './FilterType'
 
 export const createButtonList = (filterType: FilterType): HTMLUListElement => {
   const allButton = createAllButton(filterType)
