+++ src/dmm/books/add-filter-button-on-bookmark-page/createDiscountedButton.ts
@@ -1,7 +1,5 @@
-// TODO: リファクタリング
-
-import { createCurrentButton } from './createCurrentButton'
-import { createNotCurrentButton } from './createNotCurrentButton'
+import { createActiveButton } from './createActiveButton'
+import { createInactiveButton } from './createInactiveButton'
 import { createUrl } from './createUrl'
 import { FilterType } from './FilterType'
 
@@ -10,11 +8,11 @@ const TEXT = 'セール中'
 export const createDiscountedButton = (filterType: FilterType): HTMLLIElement => {
   switch (filterType) {
     case FilterType.All:
-      return createNotCurrentButton({
+      return createInactiveButton({
         text: TEXT,
         url: createUrl(FilterType.Discounted)
       })
     case FilterType.Discounted:
-      return createCurrentButton(TEXT)
+      return createActiveButton(TEXT)
   }
 }
