+++ src/dmm/books/add-filter-button-on-bookmark-page/createInactiveButton.ts
@@ -1,13 +1,17 @@
-// TODO: リファクタリング
-
-export const createCurrentButton = (text: string): HTMLLIElement => {
-  const label = document.createElement('span')
+export const createInactiveButton = ({
+  text,
+  url
+}: {
+  text: string
+  url: URL
+}): HTMLLIElement => {
+  const label = document.createElement('a')
+  label.href = url.toString()
   label.style.paddingLeft = '8px'
   label.style.paddingRight = '8px'
   label.textContent = text
 
   const button = document.createElement('li')
-  button.classList.add('current')
   button.style.width = 'auto'
   button.append(label)
   return button
