+++ src/dmm/books/add-filter-button-on-bookmark-page/main.ts
@@ -1,5 +1,3 @@
-// TODO: リファクタリング
-
 import { createFilterMenu } from './createFilterMenu'
 import {
   FilterType,
